+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- alignment of cta buttons in product lists
+
 ## [1.0.1] - 2019-10-31
 ### Added
 - Added support for PWA v6.8.0+